requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas
scikit-learn

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow
s3fs

# utilities
six>=1.14
joblib
memoized-property
termcolor

matplotlib
pydot
librosa
audioread
pydub
python-dotenv
spotipy
tensorflow
imblearn
google-cloud

uvicorn
FastApi

xgboost